Refactoring user functions

+++ src/main/java/com/bluevelvet/service/UserService.java
@@ -1,6 +1,9 @@
 package com.bluevelvet.service;
 
+import com.bluevelvet.model.Brand;
+import com.bluevelvet.model.Product;
 import com.bluevelvet.model.User;
+import com.bluevelvet.repository.RoleRepository;
 import com.bluevelvet.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +16,8 @@ public class UserService {
 
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    private RoleService roleService;
 
     public User saveUser(User user) {
         return userRepository.save(user);
@@ -26,7 +31,24 @@ public class UserService {
         return userRepository.findById(id);
     }
 
-    public void deleteUser(int id) {
-        userRepository.deleteById(id);
+    public User updateUser(int id, User user) {
+        if (userRepository.existsById(id)) {
+            user.setId(id);
+            return userRepository.save(user);
+        }
+        return null;
+    }
+
+    public boolean deleteUser(int id) {
+        if (userRepository.existsById(id)) {
+            User user = userRepository.findById(id).orElseThrow();
+            user.getRoles().clear();
+            userRepository.save(user);
+            roleService.removeUserFromAllRoles(id);
+            userRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
